+++ packages/badge-private/src/components/constants.ts
@@ -1,0 +1,4 @@
+export enum Types {
+  Info = 'info',
+  Alert = 'alert',
+}