+++ packages/filter/src/helpers/types.ts
@@ -16,7 +16,7 @@ export interface IFilterProps {
 
   value: TFilterValueType[] | string;
   onChange?: (value: TFilterValueType[], queryString: string) => void;
-  children?: (props: { badgeText: string; open?: boolean }) => JSX.Element;
+  children?: (props: { badgeNumber?: number; open?: boolean }) => JSX.Element;
 }
 
 export interface IFilterRowProps extends Partial<IFilterProps> {